chore: refactor Redux state management for the options page
