+++ src/assets/js/store.js
@@ -1,6 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit';
 
-import authReducer from '@/features/auth/authSlice';
+import authReducer from '@/features/options/optionsSlice';
 import popupReducer from '@/features/popup/popupSlice';
 
 export const store = configureStore({
